refactor(workbench): fold readWorkbenchLock into the lock claim; co-locate registry tests
acquireWorkbenchLock now returns a discriminated claim — the held lock, or who
holds it on a failed claim — so readWorkbenchLock stops being a second public
read (and a second filesystem hit on contention) and becomes internal. One
fewer export.

The registry's tests were spread across three files (registry, workbenchLock,
processLiveness) though all three exercise registry.ts. Merge them into
registry.test.ts under one shared setup and rename the real-PowerShell file to
registry.windows.test.ts. processLiveness stays an internal primitive, tested
through the registry's public API.
